+++ pkg/upload/android-proguard.go
@@ -39,6 +39,7 @@ func ProcessAndroidProguard(
 	timeout int,
 	overwrite bool,
 	dryRun bool,
+	logger log.Logger,
 ) error {
 
 	var mappingFile string
@@ -51,7 +52,7 @@ func ProcessAndroidProguard(
 			mappingPath := filepath.Join(path, "app", "build", "outputs", "mapping")
 
 			if !utils.FileExists(mappingPath) {
-				return fmt.Errorf("unable to find the mapping directory in " + path)
+				return fmt.Errorf("unable to find the mapping directory in %s", path)
 			}
 
 			if variant == "" {
@@ -72,7 +73,7 @@ func ProcessAndroidProguard(
 				appManifestPathExpected = filepath.Join(path, "app", "build", "intermediates", "merged_manifests", variant, "AndroidManifest.xml")
 				if utils.FileExists(appManifestPathExpected) {
 					appManifestPath = appManifestPathExpected
-					log.Info("Found app manifest at: " + appManifestPath)
+					logger.Info(fmt.Sprintf("Found app manifest at: %s", appManifestPath))
 				}
 			}
 
@@ -90,7 +91,7 @@ func ProcessAndroidProguard(
 							appManifestPathExpected = filepath.Join(mergedManifestPath, variant, "AndroidManifest.xml")
 							if utils.FileExists(appManifestPathExpected) {
 								appManifestPath = appManifestPathExpected
-								log.Info("Found app manifest at: " + appManifestPath)
+								logger.Info(fmt.Sprintf("Found app manifest at: %s", appManifestPath))
 							}
 						}
 					}
@@ -102,7 +103,7 @@ func ProcessAndroidProguard(
 		// Check to see if we need to read the manifest file due to missing options
 		if appManifestPath != "" && (apiKey == "" || applicationId == "" || buildUuid == "" || versionCode == "" || versionName == "") {
 
-			log.Info("Reading data from AndroidManifest.xml")
+			logger.Debug("Reading data from AndroidManifest.xml")
 			manifestData, err := android.ParseAndroidManifestXML(appManifestPath)
 
 			if err != nil {
@@ -117,7 +118,7 @@ func ProcessAndroidProguard(
 				}
 
 				if apiKey != "" {
-					log.Info("Using " + apiKey + " as API key from AndroidManifest.xml")
+					logger.Debug(fmt.Sprintf("Using %s as API key from AndroidManifest.xml", apiKey))
 				}
 			}
 
@@ -125,13 +126,13 @@ func ProcessAndroidProguard(
 				applicationId = manifestData.ApplicationId
 
 				if applicationId != "" {
-					log.Info("Using " + applicationId + " as application ID from AndroidManifest.xml")
+					logger.Debug(fmt.Sprintf("Using %s as application ID from AndroidManifest.xml", applicationId))
 				}
 			}
 
 			if noBuildUuid {
 				buildUuid = ""
-				log.Info("No build ID will be used")
+				logger.Info("No build ID will be used")
 			} else if buildUuid == "" {
 				for i := range manifestData.Application.MetaData.Name {
 					if manifestData.Application.MetaData.Name[i] == "com.bugsnag.android.BUILD_UUID" {
@@ -157,10 +158,10 @@ func ProcessAndroidProguard(
 					buildUuid = fmt.Sprintf("%x", signature)
 
 					if buildUuid != "" {
-						log.Info("Using " + buildUuid + " as build ID from classes.dex")
+						logger.Debug(fmt.Sprintf("Using %s as build ID from classes.dex", buildUuid))
 					}
 				} else {
-					log.Info("Using " + buildUuid + " as build UUID from AndroidManifest.xml")
+					logger.Debug(fmt.Sprintf("Using %s as build UUID from AndroidManifest.xml", buildUuid))
 				}
 			}
 
@@ -168,7 +169,8 @@ func ProcessAndroidProguard(
 				versionCode = manifestData.VersionCode
 
 				if versionCode != "" {
-					log.Info("Using " + versionCode + " as version code from AndroidManifest.xml")
+
+					logger.Debug(fmt.Sprintf("Using %s as version code from AndroidManifest.xml", versionCode))
 				}
 			}
 
@@ -176,12 +178,11 @@ func ProcessAndroidProguard(
 				versionName = manifestData.VersionName
 
 				if versionName != "" {
-					log.Info("Using " + versionName + " as version name from AndroidManifest.xml")
+					logger.Debug(fmt.Sprintf("Using %s as version name from AndroidManifest.xml", versionName))
 				}
 			}
 		}
-
-		log.Info("Compressing " + mappingFile)
+		logger.Info(fmt.Sprintf("Compressing %s", mappingFile))
 
 		outputFile, err := utils.GzipCompress(mappingFile)
 
@@ -198,12 +199,12 @@ func ProcessAndroidProguard(
 		fileFieldData := make(map[string]string)
 		fileFieldData["proguard"] = outputFile
 
-		err = server.ProcessFileRequest(endpoint+"/proguard", uploadOptions, fileFieldData, timeout, retries, outputFile, dryRun)
+		err = server.ProcessFileRequest(endpoint+"/proguard", uploadOptions, fileFieldData, timeout, retries, outputFile, dryRun, logger)
 
 		if err != nil {
 			if strings.Contains(err.Error(), "404 Not Found") {
-				log.Info("Trying " + endpoint)
-				err = server.ProcessFileRequest(endpoint, uploadOptions, fileFieldData, timeout, retries, outputFile, dryRun)
+				logger.Info(fmt.Sprintf("Trying %s", endpoint))
+				err = server.ProcessFileRequest(endpoint, uploadOptions, fileFieldData, timeout, retries, outputFile, dryRun, logger)
 			}
 		}