+++ pkg/upload/dart.go
@@ -38,17 +38,14 @@ func Dart(
 	overwrite bool,
 	apiKey string,
 	dryRun bool,
+	logger log.Logger,
 ) error {
-
-	log.Info("Building file list from path")
 
 	fileList, err := utils.BuildFileList(paths)
 
 	if err != nil {
-		log.Error("error building file list", 1)
+		logger.Fatal("error building file list")
 	}
-
-	log.Info("File list built")
 
 	for _, file := range fileList {
 
@@ -58,7 +55,7 @@ func Dart(
 
 		// Start processing the android symbol file
 		if isAndroidPlatform {
-			log.Info("Processing android symbol file: " + file)
+			logger.Info(fmt.Sprintf("Processing android symbol file: %s", file))
 
 			var buildId string
 			buildId, err = GetBuildIdFromElfFile(file)
@@ -72,7 +69,7 @@ func Dart(
 			fileFieldData := make(map[string]string)
 			fileFieldData["symbolFile"] = file
 
-			err := server.ProcessFileRequest(endpoint+"/dart-symbol", uploadOptions, fileFieldData, timeout, retries, file, dryRun)
+			err := server.ProcessFileRequest(endpoint+"/dart-symbol", uploadOptions, fileFieldData, timeout, retries, file, dryRun, logger)
 
 			if err != nil {
 
@@ -84,7 +81,7 @@ func Dart(
 
 		// Process iOS file
 		if isIosPlatform {
-			log.Info("Processing iOS symbol file: " + file)
+			logger.Info(fmt.Sprintf("Processing iOS symbol file: %s", file))
 
 			if iosAppPath == "" {
 				iosAppPath, err = GetIosAppPath(file)
@@ -115,7 +112,7 @@ func Dart(
 			if dryRun {
 				err = nil
 			} else {
-				err = server.ProcessFileRequest(endpoint+"/dart-symbol", uploadOptions, fileFieldData, timeout, retries, file, dryRun)
+				err = server.ProcessFileRequest(endpoint+"/dart-symbol", uploadOptions, fileFieldData, timeout, retries, file, dryRun, logger)
 			}
 
 			if err != nil {
@@ -125,7 +122,7 @@ func Dart(
 
 			continue
 		}
-		log.Info("Skipping " + file)
+		logger.Debug(fmt.Sprintf("Skipping %s - unsupported platform", file))
 	}
 
 	return nil