+++ pkg/utils/validation.go
@@ -1,12 +1,14 @@
 package utils
 
 import (
+	"fmt"
 	"os"
 	"strings"
 )
 
 type Paths []string
 type Path string
+type LogLevels string
 
 // Validate that the path(s) exist
 func (p Paths) Validate() error {
@@ -26,6 +28,7 @@ func (p Path) Validate() error {
 	return nil
 }
 
+// ContainsString Check if a string is in a slice
 func ContainsString(slice []string, target string) bool {
 	for _, element := range slice {
 		if strings.Contains(element, target) {
@@ -33,4 +36,14 @@ func ContainsString(slice []string, target string) bool {
 		}
 	}
 	return false
+}
+
+// Validate that the log level is valid
+func (l LogLevels) Validate() error {
+	switch strings.ToLower(string(l)) {
+	case "debug", "info", "warn", "fatal":
+		return nil
+	default:
+		return fmt.Errorf("invalid log level: %s", l)
+	}
 }