+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Changelog
 
-## TBD
+## 2.3.0 (2024-06-04)
+
+### Enhancements
+
+- Add the ability to set the log leveL via the `--log-level` flag [103](https://github.com/bugsnag/bugsnag-cli/pull/103)
+- Allow more flexible path searching when uploading NDK symbol files [98](https://github.com/bugsnag/bugsnag-cli/pull/98)
+
+### Fixes
+
+- Fix the error message when `--version-name` is missing [103](https://github.com/bugsnag/bugsnag-cli/pull/103)
+
+## 2.2.0 (2024-04-17)
 
 ### Enhancements
 
 - `upload android-proguard` will now attempt to automatically locate the `classes.dex` files if no build-uuid or dex-files are found or specified [92](https://github.com/bugsnag/bugsnag-cli/pull/92)
 - Added the `--no-build-uuid` option to the `upload android-*` options [92](https://github.com/bugsnag/bugsnag-cli/pull/92)
+- Added `Windows_NT` to `supported-platforms.yml` [95](https://github.com/bugsnag/bugsnag-cli/pull/95)
 
 ## 2.1.1 (2023-03-22)